requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.4
numpy==1.18.4
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.23.0
scikit-image==0.17.2
scikit-learn==0.23.0
scipy==1.4.1
selenium==3.141.0
six==1.14.0
sklearn==0.0
splinter==0.13.0
threadpoolctl==2.0.0
tifffile==2020.5.11
urllib3==1.25.9
wincertstore==0.2
